Upstream version 41.0.2272.118

+++ chrome/test/chromedriver/chrome/chrome_desktop_impl.cc
@@ -27,36 +27,36 @@
 
 namespace {
 
-bool KillProcess(base::ProcessHandle process_id) {
+bool KillProcess(base::ProcessHandle process_id, bool kill_gracefully) {
 #if defined(OS_POSIX)
-  kill(process_id, SIGKILL);
-  base::TimeTicks deadline =
-      base::TimeTicks::Now() + base::TimeDelta::FromSeconds(30);
-  while (base::TimeTicks::Now() < deadline) {
-    pid_t pid = HANDLE_EINTR(waitpid(process_id, NULL, WNOHANG));
-    if (pid == process_id)
-      return true;
-    if (pid == -1) {
-      if (errno == ECHILD) {
-        // The wait may fail with ECHILD if another process also waited for
-        // the same pid, causing the process state to get cleaned up.
+  if (!kill_gracefully) {
+    kill(process_id, SIGKILL);
+    base::TimeTicks deadline =
+        base::TimeTicks::Now() + base::TimeDelta::FromSeconds(30);
+    while (base::TimeTicks::Now() < deadline) {
+      pid_t pid = HANDLE_EINTR(waitpid(process_id, NULL, WNOHANG));
+      if (pid == process_id)
         return true;
+      if (pid == -1) {
+        if (errno == ECHILD) {
+          // The wait may fail with ECHILD if another process also waited for
+          // the same pid, causing the process state to get cleaned up.
+          return true;
+        }
+        LOG(WARNING) << "Error waiting for process " << process_id;
       }
-      LOG(WARNING) << "Error waiting for process " << process_id;
+      base::PlatformThread::Sleep(base::TimeDelta::FromMilliseconds(50));
     }
-    base::PlatformThread::Sleep(base::TimeDelta::FromMilliseconds(50));
+    return false;
   }
-  return false;
 #endif
 
-#if defined(OS_WIN)
   if (!base::KillProcess(process_id, 0, true)) {
     int exit_code;
     return base::GetTerminationStatus(process_id, &exit_code) !=
         base::TERMINATION_STATUS_STILL_RUNNING;
   }
   return true;
-#endif
 }
 
 }  // namespace
@@ -66,15 +66,15 @@ ChromeDesktopImpl::ChromeDesktopImpl(
     scoped_ptr<DevToolsClient> websocket_client,
     ScopedVector<DevToolsEventListener>& devtools_event_listeners,
     scoped_ptr<PortReservation> port_reservation,
-    base::ProcessHandle process,
-    const CommandLine& command,
+    base::Process process,
+    const base::CommandLine& command,
     base::ScopedTempDir* user_data_dir,
     base::ScopedTempDir* extension_dir)
     : ChromeImpl(http_client.Pass(),
                  websocket_client.Pass(),
                  devtools_event_listeners,
                  port_reservation.Pass()),
-      process_(process),
+      process_(process.Pass()),
       command_(command) {
   if (user_data_dir->IsValid())
     CHECK(user_data_dir_.Set(user_data_dir->Take()));
@@ -94,7 +94,6 @@ ChromeDesktopImpl::~ChromeDesktopImpl() {
       LOG(WARNING) << "chromedriver automation extension directory: "
                    << extension_dir.value();
   }
-  base::CloseProcessHandle(process_);
 }
 
 Status ChromeDesktopImpl::WaitForPageToLoad(const std::string& url,
@@ -155,8 +154,9 @@ Status ChromeDesktopImpl::GetAutomationExtension(
   return Status(kOk);
 }
 
-ChromeDesktopImpl* ChromeDesktopImpl::GetAsDesktop() {
-  return this;
+Status ChromeDesktopImpl::GetAsDesktop(ChromeDesktopImpl** desktop) {
+  *desktop = this;
+  return Status(kOk);
 }
 
 std::string ChromeDesktopImpl::GetOperatingSystemName() {
@@ -168,11 +168,16 @@ bool ChromeDesktopImpl::IsMobileEmulationEnabled() const {
 }
 
 Status ChromeDesktopImpl::QuitImpl() {
-  if (!KillProcess(process_))
+  // If the Chrome session uses a custom user data directory, try sending a
+  // SIGTERM signal before SIGKILL, so that Chrome has a chance to write
+  // everything back out to the user data directory and exit cleanly.If
+  // we're using a temporary user data directory, we're going to delete
+  // the temporary directory anyway, so just send SIGKILL immediately.
+  if (!KillProcess(process_.Handle(), !user_data_dir_.IsValid()))
     return Status(kUnknownError, "cannot kill Chrome");
   return Status(kOk);
 }
 
-const CommandLine& ChromeDesktopImpl::command() const {
+const base::CommandLine& ChromeDesktopImpl::command() const {
   return command_;
 }
